Remove unused Claude, Gemini, and Codex provider implementations
- Eliminate the Claude, Gemini, and Codex provider types and their associated methods from the MCP provider system.
- Clean up imports and reduce code complexity by removing unnecessary dependencies and functions.
- This change streamlines the provider management and focuses on the remaining active providers.
