+++ internal/mcp/providers.go
@@ -2,11 +2,7 @@ package mcp
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
 
-	"github.com/BurntSushi/toml"
 	"github.com/dosu-ai/dosu-cli/internal/config"
 )
 
@@ -47,383 +43,4 @@ func AllProviders() []Provider {
 		&MCPorterProvider{},
 		&ManualProvider{},
 	}
-}
-
-type ClaudeProvider struct{}
-
-func (p *ClaudeProvider) Name() string        { return "Claude Code" }
-func (p *ClaudeProvider) ID() string          { return "claude" }
-func (p *ClaudeProvider) SupportsLocal() bool { return true }
-func (p *ClaudeProvider) Priority() int       { return 1 }
-func (p *ClaudeProvider) DetectPaths() []string {
-	return []string{"~/.claude"}
-}
-func (p *ClaudeProvider) IsInstalled() bool {
-	return isInstalled(p.DetectPaths())
-}
-func (p *ClaudeProvider) GlobalConfigPath() string {
-	return expandHome("~/.claude.json")
-}
-
-func (p *ClaudeProvider) getConfigPath() (string, error) {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(homeDir, ".claude.json"), nil
-}
-
-func (p *ClaudeProvider) Install(cfg *config.Config, global bool) error {
-	if cfg.DeploymentID == "" {
-		return fmt.Errorf("deployment ID is required")
-	}
-
-	url := mcpURL(cfg.DeploymentID)
-
-	configPath, err := p.getConfigPath()
-	if err != nil {
-		return fmt.Errorf("failed to get claude config path: %w", err)
-	}
-
-	claudeConfig, err := loadJSONConfig(configPath)
-	if err != nil {
-		return fmt.Errorf("failed to load claude config: %w", err)
-	}
-
-	server := map[string]any{
-		"type":    "http",
-		"url":     url,
-		"headers": mcpHeaders(cfg),
-	}
-
-	if global {
-		// Add to root-level mcpServers
-		mcpServers, ok := claudeConfig["mcpServers"].(map[string]any)
-		if !ok {
-			mcpServers = make(map[string]any)
-		}
-		mcpServers["dosu"] = server
-		claudeConfig["mcpServers"] = mcpServers
-	} else {
-		// Add to projects.<cwd>.mcpServers
-		cwd, err := os.Getwd()
-		if err != nil {
-			return fmt.Errorf("failed to get current directory: %w", err)
-		}
-
-		projects, ok := claudeConfig["projects"].(map[string]any)
-		if !ok {
-			projects = make(map[string]any)
-		}
-
-		project, ok := projects[cwd].(map[string]any)
-		if !ok {
-			project = make(map[string]any)
-		}
-
-		mcpServers, ok := project["mcpServers"].(map[string]any)
-		if !ok {
-			mcpServers = make(map[string]any)
-		}
-
-		mcpServers["dosu"] = server
-		project["mcpServers"] = mcpServers
-		projects[cwd] = project
-		claudeConfig["projects"] = projects
-	}
-
-	if err := saveJSONConfig(configPath, claudeConfig); err != nil {
-		return fmt.Errorf("failed to save claude config: %w", err)
-	}
-
-	return nil
-}
-
-func (p *ClaudeProvider) IsConfigured() bool {
-	configPath, err := p.getConfigPath()
-	if err != nil {
-		return false
-	}
-	cfg, err := loadJSONConfig(configPath)
-	if err != nil {
-		return false
-	}
-	if mcpServers, ok := cfg["mcpServers"].(map[string]any); ok {
-		if _, exists := mcpServers["dosu"]; exists {
-			return true
-		}
-	}
-	return false
-}
-
-func (p *ClaudeProvider) Remove(global bool) error {
-	configPath, err := p.getConfigPath()
-	if err != nil {
-		return fmt.Errorf("failed to get claude config path: %w", err)
-	}
-
-	// Check if config exists
-	if _, err := os.Stat(configPath); os.IsNotExist(err) {
-		return nil
-	}
-
-	claudeConfig, err := loadJSONConfig(configPath)
-	if err != nil {
-		return fmt.Errorf("failed to load claude config: %w", err)
-	}
-
-	if global {
-		if mcpServers, ok := claudeConfig["mcpServers"].(map[string]any); ok {
-			delete(mcpServers, "dosu")
-		}
-	} else {
-		cwd, err := os.Getwd()
-		if err != nil {
-			return fmt.Errorf("failed to get current directory: %w", err)
-		}
-
-		if projects, ok := claudeConfig["projects"].(map[string]any); ok {
-			if project, ok := projects[cwd].(map[string]any); ok {
-				if mcpServers, ok := project["mcpServers"].(map[string]any); ok {
-					delete(mcpServers, "dosu")
-				}
-			}
-		}
-	}
-
-	if err := saveJSONConfig(configPath, claudeConfig); err != nil {
-		return fmt.Errorf("failed to save claude config: %w", err)
-	}
-
-	return nil
-}
-
-type GeminiProvider struct{}
-
-func (p *GeminiProvider) Name() string        { return "Gemini CLI" }
-func (p *GeminiProvider) ID() string          { return "gemini" }
-func (p *GeminiProvider) SupportsLocal() bool { return true }
-func (p *GeminiProvider) Priority() int       { return 7 }
-func (p *GeminiProvider) DetectPaths() []string {
-	return []string{"~/.gemini"}
-}
-func (p *GeminiProvider) IsInstalled() bool {
-	if !isInstalled(p.DetectPaths()) {
-		return false
-	}
-	_, err := exec.LookPath("gemini")
-	return err == nil
-}
-func (p *GeminiProvider) GlobalConfigPath() string {
-	return expandHome("~/.gemini/settings.json")
-}
-
-func (p *GeminiProvider) Install(cfg *config.Config, global bool) error {
-	if cfg.DeploymentID == "" {
-		return fmt.Errorf("deployment ID is required")
-	}
-
-	url := mcpURL(cfg.DeploymentID)
-
-	args := []string{"mcp", "add", "--transport", "http"}
-	if global {
-		args = append(args, "--scope", "user")
-	} else {
-		args = append(args, "--scope", "project")
-	}
-	args = append(args,
-		"--header", fmt.Sprintf("X-Dosu-API-Key: %s", cfg.APIKey),
-		"dosu",
-		url,
-	)
-
-	cmd := exec.Command("gemini", args...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("gemini mcp add failed: %w: %s", err, string(output))
-	}
-	return nil
-}
-
-func (p *GeminiProvider) IsConfigured() bool {
-	return false
-}
-
-func (p *GeminiProvider) Remove(global bool) error {
-	args := []string{"mcp", "remove"}
-	if global {
-		args = append(args, "--scope", "user")
-	} else {
-		args = append(args, "--scope", "project")
-	}
-	args = append(args, "dosu")
-
-	cmd := exec.Command("gemini", args...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("gemini mcp remove failed: %w: %s", err, string(output))
-	}
-	return nil
-}
-
-type CodexProvider struct{}
-
-func (p *CodexProvider) Name() string        { return "Codex CLI" }
-func (p *CodexProvider) ID() string          { return "codex" }
-func (p *CodexProvider) SupportsLocal() bool { return false }
-func (p *CodexProvider) Priority() int       { return 8 }
-func (p *CodexProvider) DetectPaths() []string {
-	return []string{"~/.codex"}
-}
-func (p *CodexProvider) IsInstalled() bool {
-	return isInstalled(p.DetectPaths())
-}
-func (p *CodexProvider) GlobalConfigPath() string {
-	return expandHome("~/.codex/config.toml")
-}
-
-func (p *CodexProvider) Install(cfg *config.Config, global bool) error {
-	if cfg.DeploymentID == "" {
-		return fmt.Errorf("deployment ID is required")
-	}
-
-	configPath, err := p.getConfigPath()
-	if err != nil {
-		return fmt.Errorf("failed to get codex config path: %w", err)
-	}
-
-	codexConfig, err := p.loadConfig(configPath)
-	if err != nil {
-		return fmt.Errorf("failed to load codex config: %w", err)
-	}
-
-	url := mcpURL(cfg.DeploymentID)
-
-	if codexConfig.MCPServers == nil {
-		codexConfig.MCPServers = make(map[string]CodexMCPServer)
-	}
-
-	codexConfig.MCPServers["dosu"] = CodexMCPServer{
-		URL:         url,
-		HTTPHeaders: mcpHeaders(cfg),
-	}
-
-	if err := p.saveConfig(configPath, codexConfig); err != nil {
-		return fmt.Errorf("failed to save codex config: %w", err)
-	}
-
-	return nil
-}
-
-func (p *CodexProvider) IsConfigured() bool {
-	configPath, err := p.getConfigPath()
-	if err != nil {
-		return false
-	}
-	cfg, err := p.loadConfig(configPath)
-	if err != nil {
-		return false
-	}
-	_, exists := cfg.MCPServers["dosu"]
-	return exists
-}
-
-func (p *CodexProvider) Remove(global bool) error {
-	configPath, err := p.getConfigPath()
-	if err != nil {
-		return fmt.Errorf("failed to get codex config path: %w", err)
-	}
-
-	codexConfig, err := p.loadConfig(configPath)
-	if err != nil {
-		return fmt.Errorf("failed to load codex config: %w", err)
-	}
-
-	if codexConfig.MCPServers == nil {
-		return nil
-	}
-
-	delete(codexConfig.MCPServers, "dosu")
-
-	if err := p.saveConfig(configPath, codexConfig); err != nil {
-		return fmt.Errorf("failed to save codex config: %w", err)
-	}
-
-	return nil
-}
-
-func (p *CodexProvider) getConfigPath() (string, error) {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(homeDir, ".codex", "config.toml"), nil
-}
-
-func (p *CodexProvider) loadConfig(path string) (*CodexConfig, error) {
-	cfg := &CodexConfig{}
-
-	// Check if file exists
-	if _, err := os.Stat(path); os.IsNotExist(err) {
-		// Return empty config if file doesn't exist
-		return cfg, nil
-	}
-
-	_, err := toml.DecodeFile(path, cfg)
-	if err != nil {
-		return nil, err
-	}
-	return cfg, nil
-}
-
-func (p *CodexProvider) saveConfig(path string, cfg *CodexConfig) error {
-	if err := os.MkdirAll(filepath.Dir(path), 0755); err != nil {
-		return err
-	}
-
-	f, err := os.Create(path)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	return toml.NewEncoder(f).Encode(cfg)
-}
-
-// CodexConfig represents the structure of ~/.codex/config.toml
-type CodexConfig struct {
-	MCPServers map[string]CodexMCPServer `toml:"mcp_servers"`
-	// Other fields can be added as needed
-}
-
-// CodexMCPServer represents an MCP server entry in Codex config
-type CodexMCPServer struct {
-	URL               string            `toml:"url"`
-	HTTPHeaders       map[string]string `toml:"http_headers,omitempty"`
-	BearerTokenEnvVar string            `toml:"bearer_token_env_var,omitempty"`
-	EnvHTTPHeaders    map[string]string `toml:"env_http_headers,omitempty"`
-}
-
-type ManualProvider struct{}
-
-func (p *ManualProvider) Name() string        { return "Manual Configuration" }
-func (p *ManualProvider) ID() string          { return "manual" }
-func (p *ManualProvider) SupportsLocal() bool { return false }
-
-func (p *ManualProvider) Install(cfg *config.Config, global bool) error {
-	url := mcpURL(cfg.DeploymentID)
-
-	fmt.Println("Use these details to configure the Dosu MCP server in your client:")
-	fmt.Println()
-	fmt.Printf("  Transport:      HTTP\n")
-	fmt.Printf("  Endpoint:       %s\n", url)
-	fmt.Printf("  Header:         X-Dosu-API-Key: %s\n", cfg.APIKey)
-	fmt.Println()
-
-	return nil
-}
-
-func (p *ManualProvider) Remove(global bool) error {
-	fmt.Println("\nTo remove the Dosu MCP server, manually delete the configuration from your client.")
-	return nil
 }
